отделили класс Node в отдельный файл
