+++ src/manager/Node.java
@@ -1,0 +1,50 @@
+package manager;
+
+import task.Task;
+
+import java.util.Objects;
+
+class Node {
+    public Task task;
+    public Node next;
+    public Node prev;
+
+    public Node(Node prev, Task task, Node next) {
+        this.prev = prev;
+        this.task = task;
+        this.next = next;
+    }
+
+    public Task getTask() {
+        return task;
+    }
+
+    public Node getNext() {
+        return next;
+    }
+
+    public void setNext(Node next) {
+        this.next = next;
+    }
+
+    public Node getPrev() {
+        return prev;
+    }
+
+    public void setPrev(Node prev) {
+        this.prev = prev;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Node node = (Node) o;
+        return Objects.equals(task, node.task);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(task);
+    }
+}
